requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
black==23.3.0
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.2.0
courts-db==0.10.17
cycler==0.12.1
cymem==2.0.7
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
executing==1.2.0
eyecite==2.5.2
fast-diff-match-patch==2.0.1
filelock==3.12.4
fonttools==4.46.0
frozenlist==1.3.3
fsspec==2023.6.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.20.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
huggingface-hub==0.17.2
idna==3.4
ipdb==0.13.13
ipython==8.15.0
isodate==0.6.1
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.5
langcodes==3.3.0
Levenshtein==0.21.0
lightning==2.1.2
lightning-utilities==0.10.0
lxml==4.9.2
manifest-ml==0.1.8
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
openai==0.27.8
packaging==23.1
pandas==2.2.0
pandas-stubs==2.0.2.230605
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.2
patsy==0.5.6
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.1
plotly==5.18.0
preshed==3.0.8
prompt-toolkit==3.0.39
proto-plus==1.22.2
protobuf==4.23.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pydantic_core==2.14.5
Pygments==2.16.1
pyparsing==3.0.9
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytorch-lightning==2.1.2
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==3.1.1
rdflib==6.3.2
redis==5.0.0
regcensus==0.4.5
regex==2023.6.3
reporters-db==3.2.36
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SPARQLWrapper==2.0.0
sqlitedict==2.1.0
srsly==2.4.6
sseclient-py==1.8.0
stack-data==0.6.2
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thefuzz==0.19.0
thinc==8.1.10
threadpoolctl==3.2.0
tiktoken==0.5.0
titlecase==2.4
torch==2.1.2
torchmetrics==1.2.1
tqdm==4.65.0
traitlets==5.10.1
typer==0.7.0
types-pytz==2023.3.0.0
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.6.3
tzdata==2023.3
uncertainty-calibration==0.1.4
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
word2number==1.1
xxhash==3.3.0
yarl==1.9.2
